# docs/swagger.yaml
basePath: /api/v1
definitions:
  myservice.serviceError:
    properties:
      code:
        type: integer
      error:
        type: string
    type: object
  myservice.signInEmployeeInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  myservice.signInEmployeeOutput:
    properties:
      token:
        type: string
    type: object
  myservice.signInOrgInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  myservice.signInOrgOutput:
    properties:
      token:
        type: string
    type: object
  myservice.signUpEmployeeInput:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      patronymic:
        type: string
      role_id:
        type: integer
      surname:
        type: string
    required:
    - email
    - name
    - password
    - patronymic
    - surname
    type: object
  myservice.signUpOrgInput:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - email
    - name
    - password
    type: object
info:
  contact:
    email: razmolodinivan@mail.ru
  title: POS-Ninja Backend API
  version: 0.1-alpha
paths:
  /auth/confirmCode:
    get:
      parameters:
      - description: '`org` or `employee`'
        in: query
        name: type
        type: string
      - description: 'адрес на который будет отправлено письмо (например: email@exmp.ru)'
        in: query
        name: code
        type: string
      responses:
        "200":
          description: возвращает пустой объект
          schema:
            type: object
      summary: Проверка кода подтверждения
  /auth/sendCode:
    get:
      parameters:
      - description: '`org` or `employee`'
        in: query
        name: type
        type: string
      - description: 'адрес на который будет отправлено письмо (например: email@exmp.ru)'
        in: query
        name: email
        type: string
      responses:
        "200":
          description: возвращает пустой объект
          schema:
            type: object
      summary: Отправка кода подтверждения почты
  /auth/signIn.Employee:
    post:
      consumes:
      - application/json
      description: Метод позволяет войти в аккаунт сотрудника. Работает только с токеном
        огранизации.
      parameters:
      - description: Объект для входа в огранизацию.
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/myservice.signInEmployeeInput'
      produces:
      - application/json
      responses:
        "200":
          description: Возвращает `jwt токен` при успешной авторизации
          schema:
            $ref: '#/definitions/myservice.signInEmployeeOutput'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/myservice.serviceError'
      summary: Вход для сотрудника
  /auth/signIn.Org:
    post:
      consumes:
      - application/json
      description: Метод позволяет войти в аккаунт организации.
      parameters:
      - description: Объект для входа в огранизацию.
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/myservice.signInOrgInput'
      produces:
      - application/json
      responses:
        "200":
          description: Возвращает `jwt токен` при успешной авторизации
          schema:
            $ref: '#/definitions/myservice.signInOrgOutput'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/myservice.serviceError'
      summary: Вход для организации
  /auth/signUp.Employee:
    post:
      consumes:
      - application/json
      description: Метод позволяет зарегистрировать ссотрудника. Работает только с
        токеном организации.
      parameters:
      - description: Объект для регитсрации сотрудника.
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/myservice.signUpEmployeeInput'
      produces:
      - application/json
      responses:
        "201":
          description: Возвращаемый объект при регистрации сотрудника
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/myservice.serviceError'
      summary: Регистрация сотрудника
  /auth/signUp.Org:
    post:
      consumes:
      - application/json
      description: Метод позволяет зарегистрировать организацию
      parameters:
      - description: Объект для регитсрации огранизации.
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/myservice.signUpOrgInput'
      produces:
      - application/json
      responses:
        "201":
          description: Возвращаемый объект при регистрации огранизации
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/myservice.serviceError'
      summary: Регистрация организации
swagger: "2.0"